check whether a transaction is active
to ignore the transacted and acknowlegeMode if there is an active
transaction.

The check is done using the TransactionSynchronizationRegistry object
that is looked up in JNDI using the
TransactionSynchronizationRegistryLookup config property.
By default, the property is not set and a RA will consider that a
transaction is active and will create the JMS session accordingly.

To specify the TransactionSynchronizationRegistryLookup to use, add a
config-property to ra.xml. For example

   <config-property>
      <description>The JNDI name to lookup the TransactionSynchronizationRegistry.</description>
      <config-property-name>TransactionSynchronizationRegistryLookup</config-property-name>
      <config-property-type>java.lang.String</config-property-type>
      <config-property-value>java:jboss/TransactionSynchronizationRegistry</config-property-value>
   </config-property>

Remove the checkState() method that is no longer correct since the RA
can used without active transaction (and the method in
JmsSessionFactoryImpl was not doing any checks prior to the previous
commit anyway).

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1033008