Fix InitialContext instantiation
Set the TCCL to the JmsActivation's ClassLoader to ensure that the
initial context factory can be properly instantiated and resources
looked up in a modular class loader environment.

Wrap the classloader changes in a SecurityActions to handle them when
the Security Manager is set.

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1101271